ATUALIZAÇÃO COM DEBUG - Testes novos

+++ automatização/src/main/java/com/rh_ponto/automatizacao/repository/AuditoriaPontoRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AuditoriaPontoRepository extends JpaRepository<AuditoriaPonto, Long> {
+public interface AuditoriaPontoRepository extends JpaRepository<AuditoriaPonto, Integer> {
 }
+
